Import backports during build
Just as a safety measure, make sure to import `backports` during the
build. This as much a test as a build step. Though as `conda-build`
compiles all Python files for us, it is not strictly necessary as a
build step. Still it is good to have in case `conda-build` ever changes
its behavior.